+++ CHANGELOG.md
@@ -30,6 +30,9 @@
 * foreign-toplevel: Wayland module that provides information about
   currently opened windows.
 * alsa: support for capture devices.
+* network: `ssid`, `signal`, `rx-bitrate` and `rx-bitrate` tags.
+* network: `poll-interval` option (for the new `signal` and
+  `*-bitrate` tags).
 
 
 ### Changed